+++ tomcat/src/main/java/org/apache/coyote/render/HttpRequestHandler.java
@@ -1,4 +1,4 @@
-package com.techcourse.web;
+package org.apache.coyote.render;
 
 public interface HttpRequestHandler {
     String handle(String method, String path);
